[objc-appscript] -byName: method now takes id instead of NSString* argument; renamed -[AEMTypeBase code], -[ASParserDef code] to -fourCharCode; documentation updates
git-svn-id: https://appscript.svn.sourceforge.net/svnroot/appscript@570 5e899d65-cff2-4c64-b955-3fa930e4d869

+++ objc-appscript/trunk/src/Appscript/reference.h
@@ -91,8 +91,18 @@
 - (id)getListOfType:(DescType)type;
 - (id)getListOfType:(DescType)type error:(NSError **)error;
 
+/*
+ * shortcuts for getting numerical property values as C primitives
+ */
+
+// shortcut for [[[[[ref get] requestedType: typeSInt32] returnType: typeSInt32] send] intValue]
 - (int)getIntWithError:(NSError **)error;
-- (int)getLongWithError:(NSError **)error;
+
+// shortcut for [[[[[ref get] requestedType: typeSInt32] returnType: typeSInt32] send] intValue] in 32-bit builds,
+//          and [[[[[ref get] requestedType: typeSInt64] returnType: typeSInt64] send] longValue] in 64-bit builds
+- (long)getLongWithError:(NSError **)error;
+
+// shortcut for [[[[[ref get] requestedType: typeIEEE64BitFloatingPoint] returnType: typeIEEE64BitFloatingPoint] send] doubleValue]
 - (double)getDoubleWithError:(NSError **)error;
 
 @end
